Begin refactor lecture and questions
