+++ client/src/actions/teacher_actions.js
@@ -5,6 +5,7 @@ export const TEACHER_LOGOUT = 'TEACHER_LOGOUT';
 export const TEACHER_DATA = 'TEACHER_DATA';
 export const NAVIGATE_SIDE = 'NAVIGATE_SIDE';
 export const SELECT_CLASS = 'SELECT_CLASS';
+export const ADD_LECTURE = 'ADD_LECTURE';
 
 export function teacherLogin(tId) {  
   return {
@@ -37,8 +38,17 @@ export function navigateSidebar(location) {
 }
 
 export function selectClass(newClass) {
+  console.log(newClass);
   return {
     type: SELECT_CLASS,
     payload: newClass
   };
 }
+
+// export function addLecture(lectureName) {
+
+//   return {
+//     type: SELECT_CLASS,
+//     payload: newClass
+//   };
+// }
