+++ client/src/teacherDashboard/lectures/components/InsertLecture.js
@@ -1,14 +1,17 @@
 import React, {Component} from 'react';
-import {ListGroupItem, Form, FormGroup, FormControl, ControlLabel, Button, Glyphicon} from 'react-bootstrap';
+import {ListGroupItem, ControlLabel, Button, Glyphicon} from 'react-bootstrap';
+import {Form, FormGroup, FormControl} from 'react-bootstrap';
+import { submitQuestion } from './../../../actions/teacher_actions';
 
-import './../../App.css';
+import './../../../App.css';
 
-class InsertLectureText extends Component {
+class InsertLecture extends Component {
   constructor(props) {
     super(props);
-    
-    this.state = {lectureName: ''}
-
+    this.state = {
+      lectureName: ''
+    }
+    this.componentDidMount = this.componentDidMount.bind(this);
     this.handleQuestionChange = this.handleQuestionChange.bind(this);
     this.handleKeyPress = this.handleKeyPress.bind(this);
   }
@@ -20,9 +23,26 @@ class InsertLectureText extends Component {
   handleKeyPress(e) {
     if (e.key === 'Enter') {
       e.preventDefault();
-      this.props.handleAddLecture(this.state.lectureName);
+      this.props.createLecture(this.state.lectureName);
       this.props.revertButton();
     }
+  }
+
+  handleAddLecture(testName) {
+    var newTest = {
+      testName: testName,
+      type: 'Lecture',
+      classId: this.props.classId,
+      answers: [],
+      questions: []
+    };
+    axios.post('/test/new', newTest).then(res => {
+      console.log('New Lecture sent to server!');
+    });
+    let arr = this.state.tests.slice();
+    arr.unshift(newTest);
+    this.props.addLecture(arr);
+    this.setState({tests: arr, selectedTest: null, selectedQuestions: []});
   }
 
   componentDidMount(){
@@ -46,6 +66,6 @@ class InsertLectureText extends Component {
   }
 }
 
-export default InsertLectureText;
+export default InsertLecture;
 
 //onKeyPress={(e) => this.props.handleAddLecture(e, this.state.lectureName)}
