+++ client/src/teacherDashboard/lectures/components/LectureList.js
@@ -1,10 +1,9 @@
 import React from 'react';
 import {ListGroup} from 'react-bootstrap';
 
-import LectureComponent from './LectureComponent';
+import Lecture from './Lecture';
 
-const LectureComponents = (props) => {
-  console.log(props)
+const LectureList = (props) => {
   const components = props.tests.map((test, i) => 
     <LectureComponent test={test}
       selectLectureHandler={props.selectLectureHandler}
@@ -17,4 +16,4 @@ const LectureComponents = (props) => {
   );
 };
 
-export default LectureComponents;
+export default LectureList;
