+++ client/src/teacherDashboard/lectures/components/Question.js
@@ -14,7 +14,7 @@ import {
 
 import '../../../App.css';
 
-class NewQuestion extends Component {
+class Question extends Component {
   constructor(props) {
     super(props);
 
@@ -105,4 +105,4 @@ class NewQuestion extends Component {
   }
 }
 
-export default NewQuestion;
+export default Question;
