+++ client/src/teacherDashboard/students/components/StudentPhoto.js
